feat: align about page width and font
