+++ sites/public/src/pages/about.tsx
@@ -1,12 +1,9 @@
 import React, { useEffect, useContext } from "react"
 import { PageView, pushGtmEvent, AuthContext } from "@bloom-housing/shared-helpers"
-import { t } from "@bloom-housing/ui-components"
-import Layout from "../layouts/application"
-import MaxWidthLayout from "../layouts/max-width"
+import About from "../components/content-pages/About"
 import { UserStatus } from "../lib/constants"
-import { PageHeaderLayout } from "../patterns/PageHeaderLayout"
 
-const About = () => {
+const AboutPage = () => {
   const { profile } = useContext(AuthContext)
 
   useEffect(() => {
@@ -17,25 +14,7 @@ const About = () => {
     })
   }, [profile])
 
-  return (
-    <Layout>
-      <PageHeaderLayout heading={t("pageTitle.about")} inverse>
-        <MaxWidthLayout>
-          <>
-            <p>{t("about.body1")}</p>
-            <br />
-            <p>{t("about.body2")}</p>
-            <br />
-            <p>{t("about.moreInfoContact")}</p>
-            <br />
-            <p>{t("about.thankYouPartners")}</p>
-            <br />
-            <p>{t("about.partnersList")}</p>
-          </>
-        </MaxWidthLayout>
-      </PageHeaderLayout>
-    </Layout>
-  )
+  return <About />
 }
 
-export default About
+export default AboutPage
